Finished reorganizing app structure
